Small Testing FrameWork Updated

+++ include/i386types.h
@@ -5,6 +5,6 @@
 
 #include <stdint.h>
 
-typedef enum boolean { true = 0, false = 1 } bool;
+typedef enum boolean { true = (uint8_t)0, false = (uint8_t)1 } bool;
 
 #endif // !TYPES_H
